+++ src/source/RazorEngine.Core/Templating/RazorEngineService.cs
@@ -40,10 +40,6 @@ namespace RazorEngine.Templating
             {
                 throw new ArgumentNullException("config");
             }
-            if (config.Debug && config.DisableTempFileLocking)
-            {
-                throw new InvalidOperationException("Debug && DisableTempFileLocking is not supported, you need to disable one of them. When Roslyn has been released and you are seeing this, open an issue as this might be possible now.");
-            }
 
             _config = config;
             //_core = new RazorEngineCore(config, this);